add new html-page and some methods

+++ src/main/java/org/spring/mvc_promo_acition/entiies/Prize.java
@@ -24,7 +24,7 @@ public class Prize {
 
     private String fullNameOfWinner;
 
-    private int telephoneNumberOfWinner;
+    private long telephoneNumberOfWinner;
 
     private String emailOfWinner;
